requirements.txt
docutils==0.12
future==0.15.2
lockfile==0.12.2
python-daemon==2.1.0
python-telegram-bot==3.0.0
PyYAML==3.11
requests==2.8.1
six==1.10.0
slackclient==0.16
websocket-client==0.34.0
wheel==0.24.0
